CRS-2628: Use BigDecimal for release multipliers
Change from Double to BigDecimal for release multipliers. Calculations for determinate release date and HDCED now use BigDecimal but Double is still being used for other release date types currently. They will be converted in future PRs.

Use named release multipliers in config rather than dynamic. This is to support the sentence level calculations and future breakdown where a name is more appropriate than a decimal as well as making it easier to define the appropriate level of precision for the multipliers.

One and two thirds release points are defined to 20 decimal places.

+++ src/test/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/service/HdcedCalculatorTest.kt
@@ -31,6 +31,7 @@ import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.SentenceAdjus
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.SentenceCalculation
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.StandardDeterminateSentence
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.UnadjustedReleaseDate
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.temporal.ChronoUnit
 import kotlin.math.ceil
@@ -571,7 +572,7 @@ class HdcedCalculatorTest {
     sentence.releaseDateTypes = ReleaseDateTypes(listOf(ReleaseDateType.CRD, ReleaseDateType.SLED), sentence, mock())
     val releaseDateCalculation = ReleaseDateCalculation(
       numberOfDaysToSED,
-      numberOfDaysToDeterminateReleaseDate.toDouble(),
+      BigDecimal.valueOf(numberOfDaysToDeterminateReleaseDate.toLong()),
       numberOfDaysToDeterminateReleaseDate,
       null,
     )
